Colony testing finished and refueling orders part 1
I'll construct a test of the fueling orders in a later commit.

+++ Pulsar4X/Pulsar4X.Lib/Entities/Population.cs
@@ -58,7 +58,7 @@ namespace Pulsar4X.Entities
 
         public float CivilianPopulation { get; set; }
         public float PopulationGrowthRate { get; set; }
-        public int FuelStockpile { get; set; }
+        public float FuelStockpile { get; set; }
         public float MaintenanceSupplies { get; set; }
